Update ZP60039 and ZP60040 usermods
The original version of ZP60039 and ZP60040 would cause an SD23
abend when the WTO/WTOR text is too long. The latest rework fixes
the abend by silently truncating the text.

The original jobstream contains 4 empty cards in IEAVVWTO that are
rejected by the version of IFOX00 shipped with the starter system,
they have been replaced with empty comments.

Fixes #25